Created Stocks app with EFC with database and sell and buy order implimentation, will add the print fucnyion next

+++ StocksAppWithEFC/Entities/BuyOrder.cs
@@ -24,6 +24,6 @@ namespace Entities
 
         [Required(ErrorMessage = "Stock Price is required")]
         [Range(1, 10000, ErrorMessage = "Stock Price must be between {0} and {1}")]
-        public double? stockPrice { get; set; }
+        public decimal? stockPrice { get; set; }
     }
 }
